+++ changelog.md
@@ -10,3 +10,4 @@
 
 # 1.0.3
 - Fixed a bug that would sometimes erase some of an object's groups when shifting.
+- Fixed a few crashes and incorrect values in the shift popup.